fix: jpg as default visualizer

+++ mddrt/utils/actions.py
@@ -14,21 +14,24 @@ def view_graphviz_diagram(drt_string: str, format: str):
     graph = Source(drt_string)
 
     if is_google_colab() or is_jupyter_notebook():
+        if format not in ["jpg", "png", "jpeg"]:
+            msg_error = "Format value should be a valid image extension for interactive Python Environments. Options are 'jpg', 'png' or 'jpeq'"
+            raise ValueError(msg_error)
         from IPython.display import Image, display
 
         graph_path = graph.render(filename=filename, format=file_format, cleanup=True)
         display(Image(graph_path))
     else:
-        from PIL import Image, UnidentifiedImageError
+        from PIL import Image
+
+        if format not in ["jpg", "png", "jpeg", "webp", "svg"]:
+            msg_error = "Format value should be a valid image extension for interactive Python Environments. Options are 'jpg', 'png', 'jpeq', 'webp' or 'svg'"
+            raise ValueError(msg_error)
 
         graph_path = graph.render(filename=filename, format=file_format, cleanup=True)
-        try:
-            img = Image.open(graph_path)
-            img.show()
-        except UnidentifiedImageError as e:
-            print(f"ERROR: {e}. Enter a valid image file format like png, jpg, jpeg, etc.")
-        finally:
-            os.remove(graph_path)
+        img = Image.open(graph_path)
+        img.show()
+        os.remove(graph_path)
 
 
 def is_jupyter_notebook():
